+++ lib/lockbox.rb
@@ -80,7 +80,7 @@ module Lockbox
 
   def self.encrypts_action_text_body(**options)
     ActiveSupport.on_load(:action_text_rich_text) do
-      ActionText::RichText.lockbox_encrypts :body, **options
+      ActionText::RichText.has_encrypted :body, **options
     end
   end
 end